refacto(front):Mainlayout done
Generating client with swagger
Still need to fix the ping function and the github auth

Linked 31
